+++ src/solc/pipeline.rs
@@ -18,13 +18,20 @@ pub enum Pipeline {
 
 impl Pipeline {
     ///
-    /// We always use EVMLA for Solidity <=0.7, or if the user does not want to compile via Yul.
+    /// With the upstream `solc`, we always use EVMLA for Solidity <=0.7.
+    /// With the ZKsync fork of `solc`, the default `solc` settings are used.
     ///
-    pub fn new(solc_version: &SolcVersion, force_evmla: bool) -> Self {
-        if solc_version.default < SolcCompiler::FIRST_YUL_VERSION || force_evmla {
-            Self::EVMLA
-        } else {
-            Self::Yul
+    /// Thus, `force_evmla` is used to switch to the old codegen with the upstream `solc`,
+    /// and `via_ir` is used to switch to the new codegen with the ZKsync fork of `solc`.
+    ///
+    pub fn new(solc_version: &SolcVersion, force_evmla: bool, via_ir: bool) -> Self {
+        match solc_version.l2_revision {
+            Some(_) if via_ir => Self::Yul,
+            Some(_) => Self::EVMLA,
+            None if solc_version.default < SolcCompiler::FIRST_YUL_VERSION || force_evmla => {
+                Self::EVMLA
+            }
+            None => Self::Yul,
         }
     }
 }