+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [1.5.0] - 2024-00-00
 
+### Changed
+
+- Switched to the old codegen with the ZKsync fork of `solc`
+
 ## [1.4.1] - 2024-04-24
 
 ### Added